J2N.Randomizer::Next(int): Corrected datatype of boundMinusOne from long to int

+++ src/J2N/Randomizer.cs
@@ -293,7 +293,7 @@ namespace J2N
                 return (int)((maxValue * (long)NextInt(31)) >> 31);
             }
 
-            long boundMinusOne = maxValue - 1;
+            int boundMinusOne = maxValue - 1;
             int bits, candidate;
             do
             {
